chore: better folders for router
